add loading icons to pending tests
