+++ client/src/components/TestSuite/TestRunButton/TestRunButton.tsx
@@ -11,17 +11,11 @@ import {
 
 export interface TestRunButtonProps {
   runTests: (runnableType: RunnableType, runnableId: string) => void;
-  setIsRunning: (isRunning: boolean) => void;
   runnable: TestSuite | TestGroup | Test;
   testRunInProgress: boolean;
 }
 
-const TestRunButton: FC<TestRunButtonProps> = ({
-  runTests,
-  setIsRunning,
-  runnable,
-  testRunInProgress,
-}) => {
+const TestRunButton: FC<TestRunButtonProps> = ({ runTests, runnable, testRunInProgress }) => {
   const runnableType = 'tests' in runnable ? RunnableType.TestGroup : RunnableType.TestSuite;
   const showRunButton = runnableIsTestSuite(runnable) || (runnable as TestGroup).user_runnable;
   const runButton = showRunButton ? (
@@ -33,7 +27,6 @@ const TestRunButton: FC<TestRunButtonProps> = ({
           edge="end"
           size="small"
           onClick={() => {
-            setIsRunning(true);
             runTests(runnableType, runnable.id);
           }}
           data-testid={`runButton-${runnable.id}`}
