+++ client/src/components/TestSuite/TestSuiteDetails/TestSuiteDetailsPanel.tsx
@@ -1,23 +1,53 @@
 import React, { FC } from 'react';
-import { TestGroup, RunnableType, TestSuite, Request, Result } from 'models/testSuiteModels';
+import { TestGroup, RunnableType, TestSuite, Request, Test, TestRun } from 'models/testSuiteModels';
 import DescriptionCard from './DescriptionCard';
 import TestGroupCard from './TestGroupCard';
+import TestGroupListItem from './TestGroupListItem';
+import TestListItem from './TestListItem/TestListItem';
 
 interface TestSuiteDetailsPanelProps {
-  runnable: TestSuite | TestGroup;
   runTests: (runnableType: RunnableType, runnableId: string) => void;
   updateRequest: (requestId: string, resultId: string, request: Request) => void;
-  currentTest: Result | null;
+  runnable: TestSuite | TestGroup;
+  testRun: TestRun | null;
   testRunInProgress: boolean;
 }
 
 const TestSuiteDetailsPanel: FC<TestSuiteDetailsPanelProps> = ({
-  runnable,
   runTests,
   updateRequest,
-  currentTest,
+  runnable,
+  testRun,
   testRunInProgress,
 }) => {
+  let listItems: JSX.Element[] = [];
+  if (runnable?.test_groups && runnable.test_groups.length > 0) {
+    listItems = runnable.test_groups.map((testGroup: TestGroup) => {
+      return (
+        <TestGroupListItem
+          key={`li-${testGroup.id}`}
+          runTests={runTests}
+          testRun={testRun}
+          testGroup={testGroup}
+          testRunInProgress={testRunInProgress}
+        />
+      );
+    });
+  } else if ('tests' in runnable) {
+    listItems = runnable.tests.map((test: Test) => {
+      return (
+        <TestListItem
+          key={`li-${test.id}`}
+          runTests={runTests}
+          updateRequest={updateRequest}
+          test={test}
+          testRun={testRun}
+          testRunInProgress={testRunInProgress}
+        />
+      );
+    });
+  }
+
   const descriptionCard =
     runnable.description && runnable.description.length > 0 ? (
       <DescriptionCard description={runnable.description} />
@@ -27,11 +57,12 @@ const TestSuiteDetailsPanel: FC<TestSuiteDetailsPanelProps> = ({
     <>
       <TestGroupCard
         runTests={runTests}
-        updateRequest={updateRequest}
         runnable={runnable}
-        currentTest={currentTest}
+        testRun={testRun}
         testRunInProgress={testRunInProgress}
-      />
+      >
+        {listItems}
+      </TestGroupCard>
       {descriptionCard}
     </>
   );
